+++ CHANGELOG.md
@@ -10,10 +10,13 @@
 - New shortcut to focus on the chat input box : `Shift + Esc`.
 - New shortcut to announce the last response : `Ctrl + Shift + Enter`.
 - Prepend every announce with 'PageLive' by default. This help users to identify messages from PageLive extension.
-- TODO: New shortcut to copy the last response
-- TODO: Play sound as sign that a response is being received.
-- TODO: Detect if this is new chat. If not, get the title of current chat.
+
+### On Progress
+
 - TODO: New shortcut to delete current chat / conversation.
+- TODO: Get the title of current chat, if not yet have title use default like 'no title'.
+- TODO: Play sound as sign that a response is being received.
+- TODO: New shortcut to copy the last response
 - Various bugs fix.
 
 ## v1.1.0